refactor(gateway): move search ranking loop to gateway-core (#845)
